scanning/requirements.txt
opencv-python
numpy
matplotlib
plotly
websocket-client
